ci: update github action to automatically create a version bump and changelog.
This commit introduces a dependency to conventional commits style of commit messages. The github action automatically generates a new version based on the collection of commits in this release. The factorio changelog will be a collection of commit messages in the release.
A draft release is created when the branch is not 'master'.